+++ golem-runner/src/executor.rs
@@ -290,7 +290,7 @@ mod tests {
             checked: false,
             clickable: true,
             focused: false,
-            bounds: Bounds::new(0.0, 0.0, 375.0, 812.0),
+            bounds: Bounds::new(0, 0, 375, 812),
             children: Vec::new(),
         }
     }
@@ -308,7 +308,7 @@ mod tests {
                 checked: false,
                 clickable: true,
                 focused: false,
-                bounds: Bounds::new(10.0, (i as f64) * 50.0, 200.0, 40.0),
+                bounds: Bounds::new(10, (i as i32) * 50, 200, 40),
                 children: Vec::new(),
             })
             .collect();
@@ -321,7 +321,7 @@ mod tests {
             checked: false,
             clickable: true,
             focused: false,
-            bounds: Bounds::new(0.0, 0.0, 375.0, 812.0),
+            bounds: Bounds::new(0, 0, 375, 812),
             children,
         }
     }
